resolving comments from the MR
